ml-service/requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-CORS==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
scikit-learn==1.5.0
scipy==1.13.0
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.2
Werkzeug==3.0.3
